+++ src/app/app.module.ts
@@ -27,6 +27,7 @@ import { HttpErrorInterceptor } from './core/interceptors/http-error.interceptor
 import { AppConfigService } from './core/services/app-config/app-config.service';
 import { NgxEchartsModule } from 'ngx-echarts';
 import { NotificationsButtonComponent } from './layout/top-navbar/notifications-button/notifications-button.component';
+import { CookieService } from 'ngx-cookie-service';
 
 export function storageFactory(): OAuthStorage {
     return localStorage;
@@ -161,6 +162,7 @@ renderer.link = (href, title, text) => {
         },
         { provide: OAuthStorage, useFactory: storageFactory },
         Title,
+        CookieService,
     ],
     bootstrap: [AppComponent],
 })